Adds permissions to deleting of comments

+++ mod/minds_comments/actions/minds_comments/delete.php
@@ -6,5 +6,12 @@ if(!$id){
 }
 
 $mc = new MindsComments();
-$mc->delete($type,$id);
+$comment =$mc->single($type,$id);
+
+/*
+ * @todo allow users who own the parent to delete comments
+ */
+if($comment['_source']['owner_guid']==elgg_get_logged_in_user_guid()|| elgg_is_admin_logged_in()){
+	$mc->delete($type,$id);
+}
